fix app wait error handling and logic

+++ Jenkinsfile
@@ -271,23 +271,29 @@ pipeline {
                 echo "\u001B[32mINFO: Start app ${new_app_name}\u001B[m"
                 sh "cf v3-start ${new_app_name}"
 
-                app_wait_timeout = 180
-                timeout(time: app_wait_timeout, unit: 'SECONDS') {
-                  app_ready = false
-                  while (app_ready != 'true') {
-                    app_ready = sh(script: "cf curl '/v3/apps/${new_app_guid}/processes/web/stats' | jq -r '.resources[] | select(.type=\"web\") | [inside({\"state\": \"RUNNING\"})]' | jq -sr 'add | all'", returnStdout: true).trim()
-                    echo "\u001B[32mINFO: App ${new_app_name} not ready, wait for 10 seconds...\u001B[m"
-                    sleep 10
+                catchError {
+                  app_wait_timeout = 180
+                  timeout(time: app_wait_timeout, unit: 'SECONDS') {
+                    app_ready = 'false'
+                    while (app_ready == 'false') {
+                      app_ready = sh(script: "cf curl '/v3/apps/${new_app_guid}/processes/web/stats' | jq -r '.resources[] | select(.type=\"web\") | [contains({\"state\": \"RUNNING\"})]' | jq -sr 'add | all'", returnStdout: true).trim()
+                      echo "\u001B[32mINFO: App ${new_app_name} not ready, wait for 10 seconds...\u001B[m"
+                      sleep 10
+                    }
+                    echo "\u001B[32mINFO: App ${new_app_name} is ready\u001B[m"
                   }
-                  echo "\u001B[32mINFO: App ${new_app_name} is ready\u001B[m"
                 }
 
-                echo "\u001B[32mINFO: Switching app routes\u001B[m"
-                app_routes.each {
-                  sh """
-                    cf curl '/v2/routes/${it}/apps/${new_app_guid}' -X PUT | jq -C '.'
-                    cf curl '/v2/routes/${it}/apps/${app_guid}' -X DELETE
-                  """
+                if (app_ready == 'true') {
+                  echo "\u001B[32mINFO: Switching app routes\u001B[m"
+                  app_routes.each {
+                    sh """
+                      cf curl '/v2/routes/${it}/apps/${new_app_guid}' -X PUT | jq -C '.'
+                      cf curl '/v2/routes/${it}/apps/${app_guid}' -X DELETE
+                    """
+                  }
+                } else {
+                  error "App failed to start."
                 }
 
               }
@@ -307,7 +313,7 @@ pipeline {
                   }
                   echo "\u001B[32mINFO: Cleanup old app\u001B[m"
                   sh """
-                    cf rename ${gds_app[2]} ${new_app_name}-old
+                    cf rename ${gds_app[2]} ${gds_app[2]}-delete
                     cf rename ${new_app_name} ${gds_app[2]}
                     cf curl '/v3/apps/${app_guid}' -X DELETE
                   """
